+++ src/kustology/utils/__init__.py
@@ -1,2 +1,6 @@
 # SPDX-License-Identifier: Apache-2.0
 # Copyright 2026 Eddie Allan
+
+from .walker import node_name, node_text
+
+__all__ = ["node_name", "node_text"]